Fix data decoding
Changes for hex decoding

Values that looked like they were hex encoded by the kernel were being improperly decoded by go-libaudit. For example a command line argument that is a number (e.g. 3005) or a container ID (which are hex) would be decoded and the resulting value would be treated as ASCII text (even though it was often garbage characters).

To fix the issue go-libaudit now checks if the parameter value was quoted by the kernel (e.g. a2="3005" vs a2=3005). Quoted values are strings and unquoted are hex data.

Changes for key decoding

When multiple keys are associated with audit rules (-F key=exec -F key=failure) the keys were not decoded correctly. Now multiple keys are handled and the values are returned as a slice via `Tags()` from `auparse.AuditMessage`.

Changes to execve arg handling

For EXECVE type messages the `auparse` package would create a synthetic key called `cmdline` that contained the concatenation of all the `aN` keys. Then it would remove the `aN` keys. This behavior has been removed to allow users to format the data as they desire.