Make Close safe to call more than once
Any invocations beyond the first client `Close()` are now no-ops (implemented via `sync.Once`).

This also fixes another issue. The `clearPIDOnClose` was not being used. This client will now conditionally clear the PID value on `Close()`. This shouldn't have been causing any problems because only the controlling process would be allowed to clear the PID value, but it was incorrect.