Handle ppc64 and ppc64le syscall lookups (#72)
ppc64 and ppc64le use the same syscall table as ppc. But we were missing an entry
in the syscall table for these two architectures so lookups would fail if you tried to install
an audit rule with `-F arch=ppc64le` or if go-libaudit tried to enrich messages with
the syscall names while running on those arches.