Allow file/dir/exe exclude rules
Not all kernels will support this, but Auditbeat will allow trying to set them.
I bumped the OS provided by Vagrant to test with auditd 3.0 which allows
setting the exe filter.

While comparing the binary rule representation generated by go-libaudit and
the binary rule from the kernel listing there was a 2 byte difference in the
syscall mask. I modified go-libaudit to match the kernel output. IIUC these values
shouldn't matter because they are outside the range of any existing syscall
numbers.

This was the diff I observed and accounted for.

  Diff:
  --- Expected
  +++ Actual
  @@ -17,3 +17,3 @@
    000000f0  ff ff ff ff ff ff ff ff  ff ff ff ff ff ff ff ff  |................|
  - 00000100  ff ff ff ff ff ff ff ff  ff ff 00 00 67 00 00 00  |............g...|
  + 00000100  ff ff ff ff ff ff ff ff  ff ff ff ff 67 00 00 00  |............g...|
    00000110  d2 00 00 00 00 00 00 00  00 00 00 00 00 00 00 00  |................|

Fixes #78