Revert change to Reassembler.Push from #116
Reassembler.Push has always made a copy of the provided byte slice.

It #116 it was modified to not copy the slice. But this was a breaking change
to its behavior and causes bugs for existing users. The typical use of this
function is to parse data from AuditClient.Receive which explicitly documents
that caller must make a copy of the data if they are going to retain it beyond
the next Receive call.

Relates #116