auparse: Add Owner field to AuditMessage
When using externally allocated buffers, we need to able to complete the
lifecycle when message reassembly is complete.  The reassembly callback
only provides the AuditMessage, so it's difficult to release related
buffers without tracking those separately.

Adding an Owner interface{} field to the message allows callers to
attach an object to resolve any related objects quickly.