aucoalesce/normalizations.yaml - flatten keys to avoid deep merge
Under yaml.v3 the anchors with objects do not perform a deep merge.

https://yaml.org/type/merge.html

> Keys in mapping nodes earlier in the sequence override keys specified in later mapping nodes.

Given the following, it was previously expected that the second entry
would merge to create {object: {primary: foo, secondary: bar}}, but that
does not happen under v3.

- &my_macro
  object:
    primary: foo

- <<: *my_macro
  object:
    secondary: bar

To address this, instead of using objects this change flattens the structure

The changes to normalizations.yaml were aided by
https://gist.github.com/andrewkroh/e7212a1ff033a0227f7c8d3fa7df8973.